chore(start): Neogit broken
This has several additional packages to improve the workflows.
Neogit was working ok but it stopped for some reason and I have not been
able to find the cause. This commit is to keep all the working changes
while I nuke it and add one package at a time and commit/push into main.